Moving widgetapi out of internal.
This will allow external users to develop their own widgets.
