+++ widgets/button/button.go
@@ -28,9 +28,9 @@ import (
 	"github.com/mum4k/termdash/internal/button"
 	"github.com/mum4k/termdash/internal/canvas"
 	"github.com/mum4k/termdash/internal/draw"
-	"github.com/mum4k/termdash/internal/widgetapi"
 	"github.com/mum4k/termdash/mouse"
 	"github.com/mum4k/termdash/terminal/terminalapi"
+	"github.com/mum4k/termdash/widgetapi"
 )
 
 // CallbackFn is the function called when the button is pressed.
