+++ widgets/donut/donut.go
@@ -29,8 +29,8 @@ import (
 	"github.com/mum4k/termdash/internal/draw"
 	"github.com/mum4k/termdash/internal/numbers"
 	"github.com/mum4k/termdash/internal/runewidth"
-	"github.com/mum4k/termdash/internal/widgetapi"
 	"github.com/mum4k/termdash/terminal/terminalapi"
+	"github.com/mum4k/termdash/widgetapi"
 )
 
 // progressType indicates how was the current progress provided by the caller.
