+++ widgets/sparkline/sparkline_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/mum4k/termdash/internal/draw"
 	"github.com/mum4k/termdash/internal/draw/testdraw"
 	"github.com/mum4k/termdash/internal/faketerm"
-	"github.com/mum4k/termdash/internal/widgetapi"
+	"github.com/mum4k/termdash/widgetapi"
 )
 
 func TestSparkLine(t *testing.T) {
